+++ apps/www/src/donate/PAYPAL_DONATE_URL.ts
@@ -1,0 +1,5 @@
+export const PAYPAL_HOSTED_BUTTON_ID = "9GL697FDK7ZWW"
+
+export const PAYPAL_DONATE_URL = `https://www.paypal.com/donate/?hosted_button_id=${PAYPAL_HOSTED_BUTTON_ID}`
+
+export default PAYPAL_DONATE_URL